Add XVIZLiveLoader and create base XVIZWebsocketLoader class
XVIZLiveLoader is for connecting to a live system which assumes
metadata will be missing start/end times and data will be sent
immediately upon connection.

The XVIZLiveloader will also update to the latest XVIZ message
automatically, removing the need for the playback controller.

XVIZWebsocketLoader is a base class for the common logic shared
with the XVIZStreamLoader